♻️ [REFACTOR] : 에셋 파일 이동 및 컴포넌트 파일 이동
