+++ src/assets/textEditor/bulletDot.png
