+++ src/assets/textEditor/bulletNumber.png
