+++ src/assets/textEditor/color.png
